--- app/src/main/java/co/feip/fefu2025/AnimeGenre.kt
@@ -1,8 +1,0 @@
-package co.feip.fefu2025
-
-import androidx.compose.ui.graphics.Color
-
-data class AnimeGenre(
-    val name: String,
-    val color: Color
-)
