+++ app/src/main/java/co/feip/fefu2025/domain/models/AnimeGenre.kt
@@ -1,0 +1,6 @@
+package co.feip.fefu2025.domain.models
+
+data class AnimeGenre(
+    val name: String,
+    val color: Long
+)
