+++ app/src/main/java/co/feip/fefu2025/domain/models/RatingData.kt
@@ -1,0 +1,6 @@
+package co.feip.fefu2025.domain.models
+
+data class RatingData(
+    val rating: Int,
+    val userCount: Int
+)
