+++ app/src/main/java/co/feip/fefu2025/domain/repository/AnimeRepository.kt
@@ -1,0 +1,9 @@
+package co.feip.fefu2025.domain.repository
+
+import co.feip.fefu2025.domain.models.AnimeCard
+import co.feip.fefu2025.domain.models.AnimeDetails
+
+interface AnimeRepository {
+    suspend fun getAnimeCards(): List<AnimeCard>
+    suspend fun getAnimeDetailsById(id: Int): AnimeDetails?
+}
