+++ app/src/main/java/co/feip/fefu2025/presentation/AnimeDetailScreen/AnimeDetailsState.kt
@@ -1,0 +1,9 @@
+package co.feip.fefu2025.presentation.AnimeDetailScreen
+
+import co.feip.fefu2025.domain.models.AnimeDetails
+
+data class AnimeDetailsState(
+    val details: AnimeDetails? = null,
+    val error: String? = null,
+    val isLoading: Boolean = false,
+)
